% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-epochdata.R
\docType{data}
\name{epochdata}
\alias{epochdata}
\title{Example HD-EEG epoched data}
\format{
The data frame contains five columns:
\describe{
\item{time}{Number of time point}
\item{signal}{HD-EEG signal amplitude (in microvolts)}
\item{epoch}{Trial number}
\item{electrode}{Channel label}
\item{subject}{Subject ID, 1 - representative health control subject, 2 - representative patient subject}
}
}
\source{
Central European Institute of Technology, Masaryk University, Brno, Czech Republic.
}
\usage{
epochdata
}
\description{
This dataset is a short slice of a HD-EEG dataset from visual task experiment ...
Example dataset contains amplitude values measured on 204 channels in 50 time points
The sampling frequency was 250 Hz, so the time interval between two points is 4 ms long (with 0 indicating the time of the stimulus).
The experiment was supported by Czech Health Research Council AZV NU21-04-00445.
}
\examples{
data("epochdata")

# A time series plot of all example trials for subject 1 on channel E4
  epochdata |>
    dplyr::filter(subject == 1 & electrode == "E4") |>
    ggplot(aes(time, signal, color = epoch)) +
    geom_path() +
    theme_minimal() +
    xlab("Time point") + ylab("Amplitude (microvolts)") +
    labs(color = "Trial")

}
\keyword{dataset}
