% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head_plot.R
\name{head_plot}
\alias{head_plot}
\title{Plot scalp map of EEG signal}
\usage{
head_plot(mesh, tri, mesh2d = NULL, signal = NULL, col.scale)
}
\arguments{
\item{mesh}{A data frame with coordinates of 3D point mesh.}

\item{tri}{A matrix with indices of the triangles.}

\item{mesh2d}{A data frame with coordinates of 2D point mesh (needed for triangulation computing, if argument tri is missing).}

\item{signal}{A vector with signal to plot.}

\item{col.scale}{A color scale which should be used for plotting.}
}
\value{
A rgl plot of scalp EEG signal.
}
\description{
A short description...
}
\examples{
# Tady potrebujeme doplnit priklad
}
