% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfaceplots.R
\name{interactive_surfaceplot_curves}
\alias{interactive_surfaceplot_curves}
\title{Interactive 3D surface plot of EEG signal over time}
\usage{
interactive_surfaceplot_curves(
  data,
  amplitude,
  sensor_ticks = NULL,
  col_range = NULL,
  col_scale = NULL
)
}
\arguments{
\item{data}{A data frame, tibble, or a database table containing the EEG data. Required columns are: \code{sensor}, \code{time}, and the column with the EEG amplitude specified in the argument \code{amplitude}.}

\item{amplitude}{A character string specifying the name of the column from the input data containing EEG amplitude values.}

\item{sensor_ticks}{A character vector specifying which sensor labels should be displayed on the y-axis. If \code{NULL}, no specific labels are forced, but mismatched names will throw an error if provided.}

\item{col_range}{A vector with minimum and maximum value of the amplitude used in the colour palette for plotting. If not defined, the range of interpolated signal is used.}

\item{col_scale}{Optionally, a colour scale to be utilised for plotting. If not defined, it is computed from \code{col_range}.}
}
\value{
A \code{plotly} widget object containing the interactive 3D surface plot.

Additionally, the returned object carries a \code{"diegr_metadata"} attribute with metadata.
}
\description{
Creates an interactive 3D surface plot displaying the EEG signal amplitude across different sensors and time points.
The x-axis represents time, the y-axis represents the sensors, and the z-axis (along with surface color) represents the signal amplitude.
}
\examples{
\donttest{
# Prepare data: Mean across epochs 1:13 for Subject 1 over a specific time range
edata <- pick_data(epochdata, subject_rg = 2, epoch_rg = 1:13)
data_base <- baseline_correction(edata, baseline_range = 1:9)
data_mean <- compute_mean(data_base, amplitude = "signal_base", type = "point", domain = "time")

# Selected sensors to display on the axis
selected_sensors <- c("E1", "E21", "E41", "E61", "E81", "E101",
 "E121", "E141", "E161", "E181", "E201", "E221")

# Render the interactive plot
interactive_surfaceplot_curves(data_mean, amplitude = "average", sensor_ticks = selected_sensors)
}
}
\seealso{
\code{\link{interactive_surfaceplot}}, \code{\link{plot_topo_mean}}
}
