% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{summary_stats_rt}
\alias{summary_stats_rt}
\title{Compute summary statistics of reaction times by subject}
\usage{
summary_stats_rt(data)
}
\arguments{
\item{data}{A data frame or a database table with response times dataset. Required columns: subject, epoch, RT (value of response time in ms).}
}
\value{
A data frame with summary statistics of response times by subject, consisting of the following columns:
\describe{
\item{subject}{Subject identifier.}
\item{n_epoch}{Number of epochs.}
\item{min_rt}{Minimum reaction time.}
\item{max_rt}{Maximum reaction time.}
\item{median_rt}{Median reaction time.}
\item{avg_rt}{Mean reaction time.}
\item{sd_rt}{Standard deviation of reaction time.}
}
}
\description{
Calculates basic descriptive statistics of reaction time (RT) for each
subject in the input data. Computed statistics include: the number of epochs, minimum,
maximum, median, mean, and standard deviation of RT.
}
\examples{
# Summary statistics for rtdata
summary_stats_rt(rtdata)
}
