% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animations.R
\name{animate_scalp}
\alias{animate_scalp}
\title{3D scalp plot animation in time}
\usage{
animate_scalp(
  data,
  amplitude,
  mesh,
  tri,
  coords = NULL,
  template = NULL,
  col_range = NULL,
  col_scale = NULL,
  sec = 0.3,
  frames_dir = NULL,
  output_path = NULL,
  framerate = 3,
  cleanup = TRUE
)
}
\arguments{
\item{data}{An input data frame or tibble with at least these required columns: \code{time} - the number of time point, \code{sensor} - the sensor label and the column with the EEG amplitude to plot specified in the argument \code{amplitude}.}

\item{amplitude}{A character string naming the column with EEG amplitude values.}

\item{mesh}{A \code{"mesh"} object (or a named list with the same structure) containing at least a \code{D3} element with x, y and z coordinates of a point mesh used for computing the IM model, and a \code{template} element specifying the sensor montage. If not defined, the polygon point mesh with default settings from \code{\link{point_mesh}} function is used. See details for more information about the structure.}

\item{tri}{A matrix with indices of the triangles. If missing, the triangulation is computed using \code{\link{make_triangulation}} function from \code{D2} element of the mesh.}

\item{coords}{Sensor coordinates as a tibble or data frame with named \code{x}, \code{y}, \code{z} and \code{sensor} columns. The \code{sensor} labels must match the labels in sensor column in \code{data}. If not defined, the template specified in \code{mesh$template} (or the default \code{"HCGSN256"}) is used.}

\item{template}{The kind of sensor template montage used. Available options are \code{"HCGSN256"}, \code{"biosemi128"}, \code{"biosemi256"}, and \code{"system1005"}. Default setting is \code{"HCGSN256"}.}

\item{col_range}{A vector with minimum and maximum value of the amplitude used in the colour palette for plotting. If not defined, the range of interpolated signal is used.}

\item{col_scale}{Optionally, a colour scale to be utilised for plotting. If not defined, it is computed from \code{col_range}.}

\item{sec}{The time interval used between individual animation frames, in seconds (default: 0.3).}

\item{frames_dir}{Directory where the individual frames will be saved. If NULL, the video is only displayed in viewer and the frames are not saved.}

\item{output_path}{Optional path to the output mp4 video file (".mp4" extension is required for correct rendering). If NULL, no video is created.}

\item{framerate}{Number of frames per second for the output mp4 video (default: 3).}

\item{cleanup}{Logical. Indicates, if all the PNG files should be deleted after encoding video. Default value is \code{TRUE}.}
}
\value{
If \code{frames_dir} is \code{NULL}, the function creates an interactive 3D animation in the \code{rgl} window and invisibly returns \code{NULL}.
If \code{frames_dir} is specified, individual animation frames (PNG) are saved. If \code{output_path} is also specified, an MP4 video is encoded and the function invisibly returns \code{NULL}.

Additionally, the returned object carries a \code{"diegr_metadata"} attribute with metadata such as details about the mesh used for plotting.
}
\description{
Display a topographic 3D scalp animation of the change in amplitude over time. The function enables direct rendering in Rstudio Viewer or saving the animation in MP4 format or individual frames in PNG format to the chosen location.
}
\details{
Setting the parameter \code{tri} requires defining a \code{mesh} parameter.
The parameter \code{mesh} should optimally be a \code{"mesh"} object (output from \code{\link{point_mesh}} function) or a list with the same structure (see \code{\link{point_mesh}} for more information). In that case, setting the argument \code{tri} is optional, and if it is absent, a triangulation based on the \code{D2} element of the mesh is calculated and used in the plot.
If the input \code{mesh} contains only 3D coordinates of a point mesh in \code{D3} element, the use of previously created triangulation (through \code{tri} argument) is required.

Notes:
For exporting the video, setting \code{frames_dir} together with \code{output_path} is required.

If a \code{mesh} object is provided, its internal template name (\code{mesh$template}) overrides the \code{template} argument to ensure spatial consistency.
When custom \code{coords} are provided, they are used for plotting the sensor locations. The \code{template} parameter (or \code{mesh$template}) is then used only for generating the background \code{mesh} if it is not provided.
}
\examples{
\donttest{
# This example may take a few seconds to render.
# Run only if you want to generate the full animation.
# Note: The example opens a rgl 3D viewer.
# Prepare a data structure:
s1e05 <- pick_data(epochdata, subject_rg = 1, epoch_rg = 5, time_rg = 10:20)
# Plot animation with default mesh and triangulation:
animate_scalp(s1e05, amplitude = "signal")
}
}
\seealso{
Static version: \code{\link{scalp_plot}}, animated 2D topo map: \code{\link{animate_topo}}
}
