% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-biosemi256.R
\docType{data}
\name{biosemi256}
\alias{biosemi256}
\title{Coordinates of the spherical BioSemi system with 256 electrodes}
\format{
A list with the following elements:
\describe{
\item{D2}{A tibble with 3 columns containing x and y coordinates and sensor labels (according to the BioSemi naming convention) in 2D.}
\item{D3}{A tibble with 4 columns containing x, y and z coordinates and sensor labels (according to the BioSemi naming convention) in 3D. See 'Details' for more information.}
\item{ROI}{A factor containing the name of the region to which the corresponding sensor belongs. The levels are: "central", "frontal", "occipital", "parietal", and "temporal". Sensors were assigned to anatomical regions based on their closest spatial proximity (3D Euclidean distance) to the standardized 10-05 system electrodes.}
}
}
\source{
The MNE-Python GitHub repository, \url{https://github.com/mne-tools/mne-python/tree/main/mne/channels/data/montages}
The FieldTrip GitHub repository, \url{https://github.com/fieldtrip/fieldtrip/tree/master/template/layout}
}
\usage{
data("biosemi256")
}
\description{
A dataset containing the Cartesian coordinates of high-density EEG sensor positions of the BioSemi system in 3D space on an idealized spherical surface, along with their corresponding positions in 2D space.
This template contains 256 electrode positions in the 3D space and for the 2D layout.
}
\details{
The axis orientation in the 3D case is as follows:
\itemize{
\item x-axis: left (-) to right (+),
\item y-axis: posterior (-) to anterior (+),
\item z-axis: inferior (-) to superior (+).
Because this is an idealized spherical model, the origin (0, 0, 0) represents the
geometric center of the sphere.
}

The coordinates originate from the MNE-Python (3D montage) and FieldTrip (2D layout) GitHub repositories.
}
\examples{
# A simple plot of sensor coordinates from BioSemi 256 system spherical template as points in 2D
plot(biosemi256$D2[,1:2], pch = 16, asp = 1)

}
\references{
BioSemi Headcaps: \url{https://www.biosemi.com/headcap2.htm}
}
\keyword{dataset}
