% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_rt.R
\name{boxplot_rt}
\alias{boxplot_rt}
\title{Plot interactive boxplots of response times}
\usage{
boxplot_rt(data, subject = NULL)
}
\arguments{
\item{data}{A data frame or a database table with response times dataset. Required columns: subject, epoch, RT (value of response time in ms).}

\item{subject}{A character vector with subjects to plot. If missing, boxplots are drawn for all avaliable subjects in \code{data}.}
}
\value{
A plotly object.
}
\description{
Function for plotting interactive boxplots of response time in individual epochs for selected subjects. The interactive plotly output enables to easily determine the epoch number from which outliers come and also allows to easily edit the image layout.
}
\examples{
# Display interactive boxplots for both example subjects
boxplot_rt(rtdata)
}
