% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplots.R
\name{boxplot_subject}
\alias{boxplot_subject}
\title{Plot interactive boxplots of subjects}
\usage{
boxplot_subject(
  data,
  amplitude = "signal",
  subject = NULL,
  channel,
  time_lim = NULL
)
}
\arguments{
\item{data}{A data frame or a database table with EEG dataset. Required columns: subject, sensor, time and the column with EEG amplitude named as in \code{amplitude} parameter.}

\item{amplitude}{A character specifying the name of the column from input data with an EEG amplitude values. Default is \code{"signal"}.}

\item{subject}{A vector with IDs of subjects to plot.}

\item{channel}{An integer or character ID of channel to plot.}

\item{time_lim}{A character vector with time range to plot. If not defined, the first ten time points are plotted.}
}
\value{
A plotly object.
}
\description{
Function for plotting interactive boxplots of EEG amplitude for individual subjects for selected channel within the chosen time interval. The interactive plotly output enables to easily determine the subjects with outlier amplitude and also allows to easily edit the image layout.
}
\details{
The input data frame or database table must contain at least following columns:
subject - a column with subject IDs,
sensor - a column with sensor labels,
time - a column with time point numbers,
and a column with measured EEG signal values (or their averages) called as in \code{amplitude}.
}
\examples{
# Interactive boxplots of signal from channel E34 for both subjects
# in chosen time points
## Note: it has no statistical sense to make boxplot from only 2 observations, but
## larger example dataset is not possible due to size limit of the package
boxplot_subject(epochdata, amplitude = "signal", channel = "E34", time_lim = c(10:20))

}
