% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mean.R
\name{compute_mean}
\alias{compute_mean}
\title{Calculate mean in temporal or spatial domain}
\usage{
compute_mean(
  data,
  subject = NULL,
  channel = NULL,
  group = "time",
  ex.epoch = NULL,
  time = NULL,
  base.int = NULL,
  raw = TRUE,
  type = "point"
)
}
\arguments{
\item{data}{A data frame, tibble or a database table with input data, required columns: subject, sensor, epoch, time and signal (for raw data) or signal_base (for baseline corrected data).}

\item{subject}{A vector of subject indices to be included in the average calculation. If multiple values are provided, the overall mean across all subjects will be calculated.}

\item{channel}{A character vector specifying the channels to be averaged.  If NULL, all channels present in input data are included.}

\item{group}{A character specifying the grouping factor. Default is \code{"time"} for calculation of the average at individual time points, other possibility is \code{"space"} for calculation of the average at individual space points (sensors).}

\item{ex.epoch}{Optional. A vector of epoch indices to be excluded from the average calculation.}

\item{time}{A numeric vector specifying the time points used for computing the average signal. If NULL, the whole time interval is included.}

\item{base.int}{Numeric character vector including time points to use as a baseline. See \code{\link{baseline_correction}} for details.}

\item{raw}{A logical value indicating whether the input data are in raw form (\code{raw = TRUE}) or baseline corrected (\code{raw = FALSE}). Default is \code{TRUE}.}

\item{type}{A character specifying the method of calculating the average, \code{"point"} for pointwise arithmetic average (default) and \code{"jack"} for jacknife average.}
}
\value{
A list with result average and standard deviation.
}
\description{
Function calculates the average signal for chosen subject (or average of more subjects) in temporal or spatial domain. The function computes a single overall average, considering all provided subjects and channels (or time points). If multiple subjects are provided, the result is the average across all specified subjects. Similarly, if multiple channels/time points are provided, the result is the average of all specified channels/time points.
}
\examples{
# Average signal for subject 1 and electrode E1 with baseline correction on interval 1:10
avg <- compute_mean(epochdata, subject = 1, channel = "E1", base.int = 1:10)
avg$average
}
