% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-HCGSN256.R
\docType{data}
\name{HCGSN256}
\alias{HCGSN256}
\title{Coordinates of 256-channel HCGSN sensors}
\format{
A list with following elements:
\describe{
\item{number}{Number of sensor (according to EGI GSN Technical Manual)}
\item{D2}{A tibble with 2 columns containing x and y coordinates of sensors in 2D}
\item{D3}{A tibble with 3 columns containing x, y and z coordinates of sensors in 3D}
\item{ROI}{Factor containing the name of the region to which the corresponding sensor belongs.
The levels are: "central", "frontal", "occipital", "parietal", "temporal".
The regions were determined by an expert from Central European Institute of Technology, Masaryk University, Brno, Czech Republic.}
}
}
\source{
Central European Institute of Technology, Masaryk University, Brno, Czech Republic.
}
\usage{
HCGSN256
}
\description{
A file containing coordinates ....
The coordinates belong to 256-channel HydroCel Geodesic Sensor Net template montage (averaging ... montages)
}
\examples{
data("HCGSN256")

# A simple plot of sensor coordinates as points in 2D
plot(HCGSN256$D2, pch = 16, asp = 1)

}
\references{
EGI Geodesic Sensor Net Technical Manual (2024)
}
\keyword{dataset}
