% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_waveforms.R
\name{interactive_waveforms}
\alias{interactive_waveforms}
\title{Plot interactive waveform graph}
\usage{
interactive_waveforms(data, subject, sensor, FS = 250, t0 = 251, col.palette)
}
\arguments{
\item{data}{A data frame with input data, must contain at least following columns: subject, electrode, time, signal, epoch.}

\item{subject}{A subject chosen to plot.}

\item{sensor}{An electrode to plot.}

\item{FS}{The sampling frequency. Default value is 250.}

\item{t0}{Index of the zero time point, i.e. point, where 0 ms should be marked (e.g. time of stimulus or time of response).}

\item{col.palette}{Optionally, color palette for plotting lines. If missing, the rainbow palette is used.}
}
\value{
A plotly graph.
}
\description{
Function for plotting time series of EEG signal color-coded by epoch in interactive plotly graph.
}
\examples{
# Plot waveforms for subject 1 and electrode "E65" with 250 sampling frequency rate
# and 251 zero time point
data(epochdata)
interactive_waveforms(epochdata, subject = 1, sensor = "E65", t0 = 251)
}
