% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_waveforms.R
\name{interactive_waveforms}
\alias{interactive_waveforms}
\title{Plot interactive waveform graph}
\usage{
interactive_waveforms(
  data,
  subject,
  channel,
  FS = 250,
  t0 = NULL,
  col.palette,
  base.int = NULL
)
}
\arguments{
\item{data}{A data frame, tibble or a database table with input data, must contain at least following columns: subject, sensor, time, signal, epoch.}

\item{subject}{A subject chosen to plot.}

\item{channel}{A channel to plot.}

\item{FS}{The sampling frequency. Default value is 250 Hz.}

\item{t0}{Index of the zero time point, i.e. point, where 0 ms should be marked (most often time of the stimulus or time of the response).}

\item{col.palette}{Optionally, a colour palette for plotting lines. If missing, the rainbow palette is used.}
}
\value{
A plotly graph.
}
\description{
Function for plotting time series of EEG signal colour-coded by epoch in interactive plotly graph. The output in plotly format enables to easily edit the image layout.
}
\details{
The input data frame or database table must contain at least following columns:
subject - a column with subject IDs,
sensor - a column with sensor labels,
time - a column with time point numbers,
signal - a column with measured EEG signal values,
epoch - a column with epoch numbers.
}
\examples{
# Plot waveforms for subject 1 and electrode "E65" with 250 sampling frequency rate
# and 10 as zero time point
interactive_waveforms(epochdata, subject = 1, channel = "E65", t0 = 10)

}
