% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_triangulation.R
\name{make_triangulation}
\alias{make_triangulation}
\title{Make triangulation of 2D point mesh}
\usage{
make_triangulation(mesh)
}
\arguments{
\item{mesh}{A data frame with named columns: x, y (required) and index (optionally).}
}
\value{
A three column matrix with indices of the vertices of the triangles.
}
\description{
Popsat triangulaci, prip. dolu pridat odkazy na literaturu...
}
\examples{

# Create small mesh for triangulation example
data(HCGSN256)
M <- point_mesh(dim = 2, n = 500, type = "polygon")
# Make triangulation on this mesh
TRI <- make_triangulation(M)
head(TRI)
}
