% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers_epoch.R
\name{outliers_epoch}
\alias{outliers_epoch}
\title{Select outlier epochs}
\usage{
outliers_epoch(
  data,
  subject = NULL,
  sensor = NULL,
  time = NULL,
  method,
  p = 0.975
)
}
\arguments{
\item{data}{A data frame, tibble or a database table with input data, required columns: subject, sensor, time, signal, epoch.}

\item{subject}{An integer or character ID of selected subject.}

\item{sensor}{An integer or character ID of selected sensor.}

\item{time}{A vector with time range for outliers detection. If not defined, the outliers are searching across all time points in the dataset.}

\item{method}{A character denoting the method used for outlier detection. The options are: \code{"iqr"} for interquartile range criterion, \code{"percentile"} for percentile method and \code{"hampel"} for Hampel filter method. See details for further information about methods.}

\item{p}{A probability value from \code{[0,1]} interval determining percentile to the percentile method (according to \code{probs} argument in \code{quantile()} function). The default value is set to 0.975 for the interval formed by the 2.5 and 97.5 percentiles.}
}
\value{
A list with following components:
\item{epoch_table}{A data frame with epoch ID and the number of time points in which the epoch was evaluated as outlier. (Only epochs with occurrence of outliers in at least one time point are presented.)}
\item{outliers_data}{A data frame with subset of data corresponding to the outliers found.}
}
\description{
Function for selecting outlier epochs for one subject and one sensor in chosen time points. Epochs are marked as outliers based on one of the following criteria: interquartile range criterion, percentile approach or Hampel filter method.
}
\details{
The input data frame or database table must contain at least following columns:
subject - a column with subject IDs,
sensor - a column with sensor labels,
time - a column with time point numbers,
signal - a column with measured EEG signal values,
epoch - a column with epoch numbers.

The outlier detection method is chosen through \code{method} argument. The possibilities are
\itemize{
\item \code{iqr} for the interquartile range criterion, values outside the interval \code{[lower quartile - 1.5 * IQR, upper quartile + 1.5 * IQR]}, where IQR denotes interquartile range, are considered as outliers
\item \code{percentile} for the percentile method, values outside the interval defined by the chosen percentiles are considered as outliers
\item \code{hampel} for the Hampel filter method, values outside the interval \code{[median - 3 * MAD, median + 3 * MAD]}, where MAD denotes median absolute deviation, are considered as outliers
}
}
\examples{
# Outlier epoch detection for subject 2, electrode E45 for the whole time range with IQR method
outliers_epoch(epochdata, subject = 2, sensor = "E45", method = "iqr")

# Outlier epoch detection for subject 2, electrode E45 for the whole time range
# using percentile method with 1 and 99 percentiles
outliers_epoch(epochdata, subject = 2, sensor = "E45", method = "percentile", p = 0.99)
}
