% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers_epoch.R
\name{outliers_epoch}
\alias{outliers_epoch}
\title{Select outlier epochs}
\usage{
outliers_epoch(
  data,
  subject = NULL,
  sensor = NULL,
  time = NULL,
  method,
  p = 0.975
)
}
\arguments{
\item{data}{A data frame, tibble or a database table with input data, must contain at least following columns: subject, sensor, time, signal, epoch.}

\item{subject}{An integer or character ID of selected subject.}

\item{sensor}{An integer or character ID of selected sensor.}

\item{time}{A vector with time range for outliers detection. If not defined, the outliers are searching across all time points in the dataset.}

\item{method}{A character denoting the method used for outlier detection. The options are: "iqr" for interquartile range criterion, "percentile" for percentile method and "hampel" for Hampel filter method.}

\item{p}{A probability value from \link{0,1} interval determining percentile to the percentile method (according to \code{probs} argument in \code{quantile()} function). The default value is set to 0.975 for the interval formed by the 2.5 and 97.5 percentiles.}
}
\value{
A list with following components:
\item{epoch.table}{A data frame with epoch ID and the number of time points in which the epoch was evaluated as outlier. (Only epochs with occurrence of outliers in at least one time point are presented.)}
\item{outliers.data}{A data frame with subset of data corresponding to the outliers found.}
}
\description{
Select outlier epochs
}
\examples{
data("epochdata")
# Outlier epoch detection for subject 2, electrode E45 for the whole time range with IQR method
outliers_epoch(epochdata, subject = 2, sensor = "E45", method = "iqr")

# Outlier epoch detection for subject 2, electrode E45 for the whole time range with percentile method with 1 and 99 percentiles
outliers_epoch(epochdata, subject = 2, sensor = "E45", method = "percentile", q = 0.99)
}
