% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_region.R
\name{pick_region}
\alias{pick_region}
\title{Choose region of interest}
\usage{
pick_region(
  coords = NULL,
  hemisphere = c("left", "right", "midline"),
  region = c("frontal", "central", "parietal", "occipital", "temporal"),
  ROI = NULL
)
}
\arguments{
\item{coords}{A data frame, matrix or named tibble with two columns of sensor coordinates. If not defined, HCGSN256 template is used. See details for more information about coordinate requirements.}

\item{hemisphere}{A character vector denoting hemisphere to choose. Possible values: \code{"left"}, \code{"right"}, \code{"midline"} or any combination of them. If not defined, both hemispheres with midline are chosen.}

\item{region}{A character vector denoting region to choose. Possible values: \code{"frontal"}, \code{"central"}, \code{"parietal"}, \code{"occipital"}, \code{"temporal"} or any combination of them. If not defined, all regions are chosen.}

\item{ROI}{A character vector with labels of regions. The order and length must be the same as in \code{coords}. If not defined, the predefined vector (according to HCGSN256 template determined by an expert from Central European Institute of Technology, Masaryk University, Brno, Czech Republic) is used.}
}
\value{
A subset of \code{coords} appropriate to the chosen region/hemisphere.
}
\description{
Choose region of interest
}
\details{
If the \code{coords} input is data frame or matrix with no named columns, the first column is considered as x coordinate.
For the correct selection of the hemisphere with own coordinates, it is necessary that the 2D layout is oriented with the nose up and that the midline electrodes have a zero x-coordinate.
}
\examples{

# Choosing regions from HCGSN256 template
data("HCGSN256")
# 1) temporal region in left hemisphere
pick_region(hemisphere = "left", region = "temporal")
# 2) frontal and central region
region_fc <- pick_region(region = c("frontal", "central"))
head(region_fc)

}
