% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_point_mesh.R
\name{plot_point_mesh}
\alias{plot_point_mesh}
\title{Plot point mesh}
\usage{
plot_point_mesh(
  mesh,
  sensors = T,
  names = F,
  col.sensors = "green",
  cex.sensors = 0.7,
  pch.sensors = 16,
  own.coordinates = NULL,
  col = "gray",
  pch = 20,
  cex = 0.4,
  axes = F,
  xlab = "",
  ylab = "",
  ...
)
}
\arguments{
\item{mesh}{A coordinate data frame of point mesh to plot.}

\item{sensors}{A logical value indicating whether the sensor locations should also be plotted.}

\item{names}{A logical value indicating whether the sensor names should also be plotted.}

\item{col.sensors}{The color of sensor locations points.}

\item{cex.sensors}{The cex argument for sensor location points.}

\item{pch.sensors}{The pch argument for sensor location points.}

\item{own.coordinates}{A list with coordinates of the sensor locations. If NULL, the HCGSN256 template is used.}

\item{col}{The color of mesh points. Default is gray.}

\item{pch}{The symbol used for points in the mesh.}

\item{cex}{The cex argument for points of the mesh.}

\item{axes}{A logical value indicating, if axes should be plotted. Default value is FALSE.}

\item{xlab}{The label of x axis (default is no label).}

\item{ylab}{The label of y axis (default is no label).}

\item{...}{Further graphical parameters, same as in \link{plot} function.}
}
\value{
A plot.
}
\description{
#' @description
Function for plotting a mesh of points  (not only) created by point_mesh() function.
}
\examples{
# 2D polygon point mesh with plotted sensors and default settings
M <- point_mesh(dim = 2, n = 4000, type = 'polygon')
plot_point_mesh(M, sensors = T)

# Plotting 2D circle point mesh with sensors as orange points
M <- point_mesh(dim = 2, n = 4000, type = 'circle')
plot_point_mesh(M, sensors = T, col.sensors = "orange")

}
