% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mean.R
\name{plot_topo_mean}
\alias{plot_topo_mean}
\title{Plot topographic map of average EEG signal}
\usage{
plot_topo_mean(
  data,
  mesh,
  coords = NULL,
  template = NULL,
  col_range = NULL,
  col_scale = NULL,
  contour = FALSE,
  legend = TRUE,
  names = FALSE
)
}
\arguments{
\item{data}{A data frame, tibble or a database table with input data to plot. It should be an output from \code{\link{compute_mean}} function or an object with the same structure, containing columns: \code{sensor} with sensor labels and \code{average, ci_low, ci_up} with values of average signal and lower and upper CI bounds.}

\item{mesh}{A \code{"mesh"} object (or a named list with the same structure) containing at least \code{D2} element with x and y coordinates of a point mesh used for computing IM model. If not defined, the point mesh with default settings from \code{\link{point_mesh}} function is used.}

\item{coords}{Sensor coordinates as a tibble or data frame with named \code{x}, \code{y} and \code{sensor} columns. The \code{sensor} labels must match the labels in sensor column in \code{data}. If not defined, the HCGSN256 template is used.}

\item{template}{The kind of sensor template montage used. Currently the only available option is \code{"HCGSN256"} denoting the 256-channel HydroCel Geodesic Sensor Net v.1.0, which is also a default setting.}

\item{col_range}{A vector with minimum and maximum value of the amplitude used in the colour palette for plotting. If not defined, the range of input data (average and CI bounds) is used.}

\item{col_scale}{Optionally, a colour scale to be utilised for plotting. If not defined, it is computed from \code{col_range}.}

\item{contour}{Logical. Indicates, whether contours should be plotted in the graph. Default value is \code{FALSE}.}

\item{legend}{Logical. Indicates, whether legend should be displayed beside the graph. Default value is \code{TRUE}.}

\item{names}{A logical value indicating whether the sensor names should also be plotted (default value is \code{FALSE}).}
}
\value{
A \code{ggplot} object showing the topographic map of the signal for chosen time.
}
\description{
Plot a topographic circle or polygon map of the average EEG signal amplitude using topographic colour scale. The function displays topographic map of average and lower and upper confidence interval bounds.
The thin-plate spline interpolation model \eqn{\text{IM:}\; \mathbb{R}^2 \rightarrow \mathbb{R}} is used for signal interpolation between the sensor locations.
The output in the form of a \code{ggplot} object allows to easily edit the result image properties. For animated version of the plot see \code{\link{animate_topo_mean}}.
}
\examples{
# Plot average topographic map with CI bounds of signal for subject 2 from the time point 10
# (the time of the stimulus) without the outliers (epoch 14 and 15)

# a) preparing data
# a1) extract required data
edata <- epochdata |>
dplyr::filter(subject == 2 & time \%in\% 1:10 & epoch \%in\% 1:13)
# a2) baseline correction (needed for suitable topographic map)
data_base <- baseline_correction(edata, base_int = 1:10)
# a3) average computing
data_mean <- compute_mean(data_base, amplitude = "signal_base", subject = 2, time = 10,
 type = "jack", group = "space")


# b) plotting the topographic map with legend
plot_topo_mean(data = data_mean, template = "HCGSN256", legend = TRUE)
}
\seealso{
\code{\link{animate_topo_mean}, \link{topo_plot}}
}
