% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_mesh.R
\name{point_mesh}
\alias{point_mesh}
\title{Create regular mesh of points}
\usage{
point_mesh(dim, n = 10000, r, template = "HCGSN256", type = "circle")
}
\arguments{
\item{dim}{A number indicating a dimension of the mesh: 2 for two dimensional and 3 for three dimensional mesh.}

\item{n}{Optionally, the required number of mesh points. Default setting is n = 10 000. (doplnit, ze cislo neni presne)}

\item{r}{Optionally, desired radius of a circular mesh. If not defined, it is selected according to the sensor locations. (doplnit, kde je nulovy bod)}

\item{template}{The kind of sensor template montage used. Default setting 'HCGSN256' denotes the 256-channel HydroCel Geodesic Sensor Net v.1.0.}

\item{type}{A character indicating the shape of the mesh with 2 possible values: 'circle' for circular mesh, 'polygon' for irregular polygon shape with boundaries defined by used sensor locations.}
}
\value{
A data frame with number of columns according to selected dimension of coordinates of mesh points.
}
\description{
Function creates a regular (in the sense ...) mesh of points on the space defined by sensor coordinates.
}
\examples{
# Computing circle 2D mesh with starting number 4000 points for HCGSN256 template
M <- point_mesh(dim = 2, n = 4000, type = 'circle')
# Computing polygon 3D mesh with starting number 2000 points for HCGSN256 template
M <- point_mesh(dim = 3, n = 2000, type = 'polygon')
}
