% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_mesh.R
\name{point_mesh}
\alias{point_mesh}
\title{Create regular mesh of points}
\usage{
point_mesh(dim, n = 10000, r, template = "HCGSN256", type = "circle")
}
\arguments{
\item{dim}{A number (or a vector) indicating a dimension of the mesh: 2 for two dimensional, 3 for three dimensional mesh and c(2,3) for both of them in one output.}

\item{n}{Optionally, the required number of mesh points. Default setting is n = 10 000.}

\item{r}{Optionally, desired radius of a circular mesh. If not defined, it is selected according to the sensor locations. (The (0,0) point corresponds to a reference electrode located at the vertex.)}

\item{template}{The kind of sensor template montage used. Default setting 'HCGSN256' denotes the 256-channel HydroCel Geodesic Sensor Net v.1.0.}

\item{type}{A character indicating the shape of the mesh with 2 possible values: 'circle' for circular mesh, 'polygon' for irregular polygon shape with boundaries defined by used sensor locations.}
}
\value{
Returns an object of class \code{"mesh"}. It is a data frame with mesh coordinates - number of columns corresponds to the selected dimension or a list containing the following components:

\item{D2}{A data frame with x and y coordinates of the created point mesh.}
\item{D3}{A data frame with x, y and z coordinates of the created point mesh.}
\item{template}{A character indicating the template of the sensor coordinates used for mesh computing.}
}
\description{
Function creates a data.frame (or list of data.frames) with coordinates of a regular (in the sense of the equidistant distance between mesh nodes) mesh of points on the space defined by sensor coordinates.
}
\details{
The number \code{n} for controlling the mesh density is only approximate value. The final number of mesh nodes depends on the exact shape of the polygon (created as a convex hull of the sensor locations), and is only close to, not exactly equal to, the number \code{n}.
}
\examples{
# Computing circle 2D mesh with starting number 4000 points for HCGSN256 template
M <- point_mesh(dim = 2, n = 4000, type = 'circle')

# Computing polygon 3D mesh with starting number 2000 points for HCGSN256 template
M <- point_mesh(dim = 3, n = 2000, type = 'polygon')

# Computing coordinates of a polygon mesh in 2D and 3D in one step:
M <- point_mesh(dim = c(2,3), n = 2000, type = 'polygon')
}
\references{
EGI Geodesic Sensor Net Technical Manual (2024)
}
