% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{summary_stats_rt}
\alias{summary_stats_rt}
\title{Compute summary statistics of reaction times by subject}
\usage{
summary_stats_rt(data)
}
\arguments{
\item{data}{A data frame or a database table with response times dataset. Required columns: \code{subject}, \code{epoch}, \code{RT} (value of response time in ms).
Optional column: \code{condition} for computing summary statistics per condition.}
}
\value{
A tibble with summary statistics of response times by subject and, if present, by condition, consisting of the following columns:
\describe{
\item{subject}{Subject identifier.}
\item{condition}{Experimental condition (only if present in the input data).}
\item{n_epoch}{Number of epochs.}
\item{min_rt}{Minimum reaction time.}
\item{max_rt}{Maximum reaction time.}
\item{median_rt}{Median reaction time.}
\item{avg_rt}{Mean reaction time.}
\item{sd_rt}{Standard deviation of reaction time.}
}
}
\description{
Calculates basic descriptive statistics of reaction time (RT) for each
subject in the input data. If condition is present, results are computed per subject and condition. Computed statistics include: the number of epochs, minimum,
maximum, median, mean, and standard deviation of RT.
}
\examples{
# Summary statistics for rtdata (two different subjects, no conditions)
summary_stats_rt(rtdata)

# Summary statistics for data with conditions
# 1. create data
data_cond <- rtdata
data_cond$condition <- c(rep("a", 7), rep("b", 7), rep("a", 8), rep("b",7))
# 2. compute statistics
summary_stats_rt(data_cond)
}
