% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-system1005.R
\docType{data}
\name{system1005}
\alias{system1005}
\title{Coordinates of the spherical 10-05 system sensors}
\format{
A list with the following elements:
\describe{
\item{D2}{A tibble with 3 columns containing x and y coordinates and sensor labels (according to the standard 10-05 naming convention) in 2D.}
\item{D3}{A tibble with 4 columns containing x, y and z coordinates and sensor labels (according to the standard 10-05 naming convention) in 3D. See 'Details' for more information.}
\item{ROI}{A factor containing the name of the region to which the corresponding sensor belongs. The levels are: "central", "frontal", "occipital", "parietal", and "temporal". See Details for more information about assigning sensors to locations.}
}
}
\source{
The MNE-Python GitHub repository, \url{https://github.com/mne-tools/mne-python/tree/main/mne/channels/data/montages}
}
\usage{
data("system1005")
}
\description{
A dataset containing the Cartesian coordinates of high-density EEG sensor positions of the 10-05 system in 3D space on an idealized spherical surface, along with their corresponding positions in 2D space.
This template contains 335 matched electrode positions for both the 3D space and the 2D layout.
Fiducial markers and extremely low electrodes (such as those on the lower neck and cheeks) originally present in the full 10-05 system are excluded from both dimensions to maintain visual clarity and structural consistency in topographical modeling.
}
\details{
The axis orientation in the 3D case is as follows:
\itemize{
\item x-axis: left (-) to right (+),
\item y-axis: posterior (-) to anterior (+),
\item z-axis: inferior (-) to superior (+).
Because this is an idealized spherical model, the origin (0, 0, 0) represents the
geometric center of the sphere. The electrodes are symmetrically distributed across
this perfect spherical surface and the electrode Cz is located exactly at
the apex of the sphere at coordinates (0, 0, 1).
}

The electrodes are mapped to anatomical regions (\code{ROI}) based on their name prefixes. Boundary electrodes are assigned to their closest primary region (e.g., 'FC' and 'FCC' are grouped under Central, while 'FFC' is Frontal; 'PO' is assigned to Occipital, and 'CP' to Parietal).

The coordinates originate from the MNE-Python GitHub repository.
}
\examples{
# A simple plot of sensor coordinates from 10-05 system spherical template as points in 2D
plot(system1005$D2[,1:2], pch = 16, asp = 1)

}
\references{
Oostenveld, R., Praamstra, P. (2001). The five percent electrode system
for high-resolution EEG and ERP measurements. \emph{Clinical Neurophysiology},
112(4), 713-719. \doi{10.1016/s1388-2457(00)00527-7}
}
\keyword{dataset}
