% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topo_plot.R
\name{topo_plot}
\alias{topo_plot}
\title{Plot topographic map of EEG signal}
\usage{
topo_plot(data, mesh, coords = HCGSN256$D2, col.range = NULL, col.scale = NULL)
}
\arguments{
\item{data}{A vector with signal to plot.}

\item{mesh}{A data frame with x and y coordinates of a point mesh used for computing IM model. If not defined, the point mesh with default settings from point_mesh() function is used.}

\item{coords}{Sensor coordinates. If not defined, the HCGSN256 template is used.}

\item{col.range}{A vector ... See details.}

\item{col.scale}{A color scale which should be used for plotting. If not defined, it is computed from col.range.}
}
\value{
A plot.
}
\description{
Doplnit podrobnosti o skale atd.
}
\examples{
data(HCGSN256)
data(epochdata)
# Plot average topographic map of signal for subject 1 from the time point 251
# the outliers - epoch 47 and 48 are extracted before computing average
data <- epochdata |>
dplyr::filter(time == 251 & subject == 1 & !epoch \%in\% c(47,48)) |>
dplyr::select(signal, electrode, epoch) |>
dplyr::group_by(electrode) |>
dplyr::mutate(average = mean(signal, na.rm = TRUE))

data <- data$average[1:204]
topo_plot(data = data, col.range = c(-40, 40))

}
